# swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AyuTrace Backend API",
    "version": "1.0.0",
    "description": "A comprehensive API for managing ayurvedic herbs supply chain traceability system",
    "contact": {
      "name": "AyuTrace Team",
      "email": "support@ayutrace.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.ayutrace.com",
      "description": "Production server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token obtained from login endpoint"
      }
    },
    "schemas": {
...
      "RawMaterialBatchStatus": {
        "type": "string",
        "enum": [
          "CREATED",
          "IN_PROCESSING",
          "PROCESSED",
          "QUARANTINED"
        ],
        "description": "Status of raw material batch in the processing pipeline",
        "example": "CREATED"
      },
      "SupplyChainEventType": {
        "type": "string",
        "enum": [
          "PROCESSING",
          "TESTING",
          "TRANSFER",
          "STORAGE",
          "PACKAGING"
        ],
        "description": "Type of supply chain event",
        "example": "PROCESSING"
      },
      "FinishedGoodProductType": {
        "type": "string",
        "enum": [
          "POWDER",
          "CAPSULE",
          "TABLET",
          "SYRUP",
          "OIL",
          "CREAM"
        ],
        "description": "Type of finished ayurvedic product",
        "example": "POWDER"
      },
...
      "QREntityType": {
        "type": "string",
        "enum": [
          "RAW_MATERIAL_BATCH",
          "FINISHED_GOOD",
          "SUPPLY_CHAIN_EVENT",
          "LAB_TEST",
          "CERTIFICATE"
        ],
        "description": "Type of entity that the QR code represents",
        "example": "RAW_MATERIAL_BATCH"
      },
...
      "QuantityUnit": {
        "type": "string",
        "enum": [
          "KG",
          "TONNES",
          "GRAMS",
          "POUNDS",
          "PIECES",
          "BOTTLES",
          "BOXES"
        ],
        "description": "Unit of measurement for quantities",
        "example": "KG"
      },
...
    }
  },
  "paths": {
...
    "/api/collections": {
      "post": {
        "summary": "Create a new collection event",
        "description": "Record a new herb collection event with quality metrics and location data",
        "tags": [
          "Collection Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "quantityKg",
                  "initialQualityMetrics",
                  "location"
                ],
                "properties": {
                  "herbSpeciesId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Species UUID (optional)",
                    "example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
                  },
                  "quantityKg": {
                    "type": "number",
                    "format": "float",
                    "minimum": 0.1,
                    "description": "Quantity of herbs collected in kilograms",
                    "example": 15.5
                  },
                  "initialQualityMetrics": {
                    "type": "string",
                    "description": "JSON string containing quality assessment data",
                    "example": "{\"moisture\": 12.5, \"purity\": 95, \"color\": \"green\", \"aroma\": \"strong\"}"
                  },
                  "photoUrl": {
                    "type": "string",
                    "format": "uri",
                    "description": "URL of collection photo (optional)",
                    "example": "https://example.com/collection-photo.jpg"
                  },
                  "location": {
                    "type": "string",
                    "description": "JSON string with GPS coordinates",
                    "example": "{\"latitude\": 17.6868, \"longitude\": 74.0183}"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Collection event recorded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Collection event recorded successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "eventId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "collectorId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "collectionDate": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "quantity": {
                          "type": "number"
                        },
                        "location": {
                          "type": "string"
                        },
                        "latitude": {
                          "type": "number"
                        },
                        "longitude": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/raw-material-batches": {
      "post": {
        "summary": "Create a new raw material batch",
        "description": "Creates a new raw material batch for tracking herbs through the supply chain",
        "tags": [
          "Raw Material Batches"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "herbName",
                  "quantity",
                  "unit"
                ],
                "properties": {
                  "herbName": {
                    "type": "string",
                    "description": "Name of the herb",
                    "example": "Ashwagandha"
                  },
                  "scientificName": {
                    "type": "string",
                    "description": "Scientific name of the herb",
                    "example": "Withania somnifera"
                  },
                  "quantity": {
                    "type": "number",
                    "description": "Quantity of the batch",
                    "example": 100.5
                  },
                  "unit": {
                    "$ref": "#/components/schemas/QuantityUnit"
                  },
                  "description": {
                    "type": "string",
                    "description": "Additional description",
                    "example": "High quality organic ashwagandha roots"
                  },
                  "notes": {
                    "type": "string",
                    "description": "Any additional notes",
                    "example": "Harvested during optimal season"
                  },
                  "currentOwnerId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "ID of the current owner",
                    "example": "b1c2d3e4-f5g6-7890-1234-567890abcdef"
                  },
                  "collectionEventIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "description": "Array of collection event IDs associated with this batch",
                    "example": [
                      "a1b2c3d4-e5f6-7890-1234-567890abcdef"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Raw material batch created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "summary": "Get all raw material batches",
        "description": "Retrieve a list of raw material batches with pagination and filtering",
        "tags": [
          "Raw Material Batches"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of items per page"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "$ref": "#/components/schemas/RawMaterialBatchStatus"
            },
            "description": "Filter by batch status"
          },
          {
            "in": "query",
            "name": "herbName",
            "schema": {
              "type": "string"
            },
            "description": "Filter by herb name"
          }
        ],
        "responses": {
          "200": {
            "description": "List of raw material batches"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/supply-chain": {
      "post": {
        "summary": "Create a new supply chain event",
        "description": "Records a new event in the supply chain tracking system",
        "tags": [
          "Supply Chain"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "eventType",
                  "handlerId",
                  "fromLocationId",
                  "toLocationId"
                ],
                "properties": {
                  "eventType": {
                    "$ref": "#/components/schemas/SupplyChainEventType"
                  },
                  "timestamp": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Event timestamp (defaults to current time)",
                    "example": "2025-09-05T10:30:00Z"
                  },
                  "rawMaterialBatchId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Associated raw material batch ID (optional)",
                    "example": "c1d2e3f4-g5h6-7890-1234-567890abcdef"
                  },
                  "finishedGoodId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Associated finished good ID (optional)",
                    "example": "e1f2g3h4-i5j6-7890-1234-567890abcdef"
                  },
                  "handlerId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "ID of the user handling the event",
                    "example": "b1c2d3e4-f5g6-7890-1234-567890abcdef"
                  },
                  "fromLocationId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "ID of the source organization/location",
                    "example": "f1g2h3i4-j5k6-7890-1234-567890abcdef"
                  },
                  "toLocationId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "ID of the destination organization/location",
                    "example": "g1h2i3j4-k5l6-7890-1234-567890abcdef"
                  },
                  "notes": {
                    "type": "string",
                    "description": "Additional notes about the event",
                    "example": "Quality check passed, transferred to processing facility"
                  },
                  "custody": {
                    "type": "object",
                    "description": "Custody information as JSON",
                    "example": {
                      "temperature": "20°C",
                      "humidity": "45%",
                      "transportMethod": "refrigerated truck"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Supply chain event created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "summary": "Get all supply chain events",
        "description": "Retrieve supply chain events with pagination and filtering",
        "tags": [
          "Supply Chain"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of items per page"
          },
          {
            "in": "query",
            "name": "eventType",
            "schema": {
              "$ref": "#/components/schemas/SupplyChainEventType"
            },
            "description": "Filter by event type"
          },
          {
            "in": "query",
            "name": "batchId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter by batch ID"
          },
          {
            "in": "query",
            "name": "handlerId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter by handler ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of supply chain events"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/finished-goods": {
      "post": {
        "summary": "Create a new finished good",
        "description": "Creates a new finished product with its composition",
        "tags": [
          "Finished Goods"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "productName",
                  "productType",
                  "quantity",
                  "unit",
                  "composition"
                ],
                "properties": {
                  "productName": {
                    "type": "string",
                    "description": "Name of the finished product",
                    "example": "Ashwagandha Churna"
                  },
                  "productType": {
                    "$ref": "#/components/schemas/FinishedGoodProductType"
                  },
                  "description": {
                    "type": "string",
                    "description": "Product description",
                    "example": "Pure Ashwagandha powder for wellness"
                  },
                  "quantity": {
                    "type": "number",
                    "description": "Quantity produced",
                    "example": 50
                  },
                  "unit": {
                    "$ref": "#/components/schemas/QuantityUnit"
                  },
                  "batchNumber": {
                    "type": "string",
                    "description": "Batch number for the product",
                    "example": "FG-ASH-2025-001"
                  },
                  "expiryDate": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Product expiry date",
                    "example": "2027-09-05T00:00:00Z"
                  },
                  "manufacturerId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "ID of the manufacturer",
                    "example": "b1c2d3e4-f5g6-7890-1234-567890abcdef"
                  },
                  "composition": {
                    "type": "array",
                    "description": "Composition of the finished good",
                    "minItems": 1,
                    "items": {
                      "type": "object",
                      "required": [
                        "rawMaterialBatchId",
                        "percentage",
                        "quantityUsed"
                      ],
                      "properties": {
                        "rawMaterialBatchId": {
                          "type": "string",
                          "description": "Raw material batch ID",
                          "example": "4cbcc605-7b7d-4bbc-975a-72cf68b84776"
                        },
                        "percentage": {
                          "type": "number",
                          "description": "Percentage of this raw material in the finished good",
                          "example": 100
                        },
                        "quantityUsed": {
                          "type": "number",
                          "description": "Quantity of raw material used",
                          "example": 50
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Finished good created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "summary": "Get all finished goods",
        "description": "Retrieve finished goods with pagination and filtering",
        "tags": [
          "Finished Goods"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of items per page"
          },
          {
            "in": "query",
            "name": "productType",
            "schema": {
              "$ref": "#/components/schemas/FinishedGoodProductType"
            },
            "description": "Filter by product type"
          },
          {
            "in": "query",
            "name": "productName",
            "schema": {
              "type": "string"
            },
            "description": "Filter by product name"
          },
          {
            "in": "query",
            "name": "manufacturerId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter by manufacturer ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of finished goods"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/documents": {
      "post": {
        "summary": "Upload a new document and link to an entity",
        "description": "Upload a document file and associate it with a specific entity (batch, event, etc.)",
        "tags": [
          "Documents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file",
                  "documentType",
                  "entityType",
                  "entityId"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Document file to upload (max 10MB)"
                  },
                  "documentType": {
                    "type": "string",
                    "enum": [
                      "CERTIFICATE",
                      "PHOTO",
                      "INVOICE",
                      "REPORT",
                      "TEST_RESULT",
                      "LICENSE",
                      "OTHER"
                    ],
                    "description": "Type of document being uploaded"
                  },
                  "entityType": {
                    "type": "string",
                    "enum": [
                      "COLLECTION_EVENT",
                      "RAW_MATERIAL_BATCH",
                      "SUPPLY_CHAIN_EVENT",
                      "FINISHED_GOOD"
                    ],
                    "description": "Type of entity to link the document to (required)",
                    "example": "RAW_MATERIAL_BATCH"
                  },
                  "entityId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "UUID of the entity to link the document to (required)",
                    "example": "a1b2c3d4-e5f6-7890-abcd-1234567890ab"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the document",
                    "example": "Organic certification for ashwagandha batch #001"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document uploaded successfully"
          },
          "400": {
            "description": "Validation error or file upload error"
          },
          "401": {
            "description": "Unauthorized (missing or invalid token)"
          }
        }
      },
      "get": {
        "summary": "Get all documents",
        "description": "Retrieve documents with pagination and filtering",
        "tags": [
          "Documents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of items per page"
          },
          {
            "in": "query",
            "name": "documentType",
            "schema": {
              "type": "string"
            },
            "description": "Filter by document type"
          }
        ],
        "responses": {
          "200": {
            "description": "List of documents"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/qr-codes": {
      "post": {
        "summary": "Generate a new QR code",
        "description": "Generate a QR code for traceability of batches or finished goods",
        "tags": [
          "QR Codes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "entityType",
                  "entityId"
                ],
                "properties": {
                  "entityType": {
                    "$ref": "#/components/schemas/QREntityType"
                  },
                  "entityId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "ID of the entity to generate QR code for",
                    "example": "e1f2g3h4-i5j6-7890-1234-567890abcdef"
                  },
                  "purpose": {
                    "type": "string",
                    "description": "Purpose of the QR code",
                    "example": "Product traceability and authentication"
                  },
                  "customData": {
                    "type": "object",
                    "description": "Additional custom data for the QR code",
                    "example": {
                      "batchInfo": "Premium Ashwagandha",
                      "harvestDate": "2025-08-15"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "QR code generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "qrCodeId": {
                          "type": "string",
                          "description": "Unique QR code ID"
                        },
                        "entityType": {
                          "$ref": "#/components/schemas/QREntityType"
                        },
                        "entityId": {
                          "type": "string"
                        },
                        "qrHash": {
                          "type": "string",
                          "description": "Unique hash for the QR code"
                        },
                        "qrImage": {
                          "type": "string",
                          "format": "data-url",
                          "description": "QR code image as a Data URL"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "QR code generated successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "summary": "Get all QR codes",
        "description": "Retrieve QR codes with pagination and filtering",
        "tags": [
          "QR Codes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of items per page"
          },
          {
            "in": "query",
            "name": "entityType",
            "schema": {
              "$ref": "#/components/schemas/QREntityType"
            },
            "description": "Filter by entity type"
          },
          {
            "in": "query",
            "name": "entityId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter by entity ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of QR codes"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}